WIP - Begin to refactor so that activities are stored on visits
After a bit more discussion with Andrew, we realized that we could
probably both save a pile of database space _and_ simplify the code base
quite a bit if we shift activities to a serialized field on the visits
table.

We never retrieve activities except as part of a visit, plus this allows
us the opportunity to work with the Rails 5/6 Attributes API which is
a pretty powerful way to work with structure data in the database in a
domain-driven manner.

This is a chainsaw refactor, so I don't expect it to land, but I would
like to talk through what I learned  when spiking it out, and how that may
(or may not) play into how we build out this service going forward.