Begin removing user_id from activities [sh24669]
For some reason, we decided not only do we want to store the user id in
both the visits and the activities table; we *also* want to index the
user id in both places.

This results in... well; a lot of wasted space because we store the
user_id twice.

Dumping this index should give us back 50GB+ of data; and dumping the
*value* will probably save us 25GB+.

After which, I think we can safely drop WhoWasHere to a S1 database;
especially if we add automatically delete old records on a quarterlyish
basis.