Upgrade to latest Ruby [sh31727] (#110)
* Upgrade to latest Ruby [sh31727]

https://app.shortcut.com/transparentclassroom/story/31727

This is probably not 100% necessary, but I always get skeevy when we're
on onlder versions of stuff.

* Install additional dependencies to fix failed startup [sh31727]

Rails 6 and Ruby 3.4 do not play particularly nice with each other.

I think it will be good to do a Rails 6 to 7 to 8 upgrade, but I am not
planning to bundle that in this set of changes...